+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## [[v0.3.0]](https://github.com/mlange-42/arche-model/compare/v0.2.0...v0.3.0)
 
+### Breaking changes
+
+* Methods `observer.Grid.X` and `observer.Grid.Y` take an `int` argument and return one value instead of all (#38)
+
 ### Features
 
 * Observer `RowToTable` as adapter from `Row` to `Table` observer (#37)
+* Observer `MatrixToGrid` as adapter from `Matrix` to `Grid` observer (#38)
 * System `PerfTimer` prints total step and average time per tick on finalization (#37)
 
 ## [[v0.2.0]](https://github.com/mlange-42/arche-model/compare/v0.1.0...v0.2.0)